Pluggable crypto and environment
