+++ Cargo.toml
@@ -8,8 +8,8 @@ description = "A Rust implementation of deserializing&decrypting RedStone payloa
 [features]
 default = ["pure"]
 
-# Pure Rust, no specific network extension
-pure = ["primitive-types"]
+# Pure Rust, no specific network extension. Uses k256 for cryptographic
+pure = ["k256"]
 
 # An extension for Casper network
 network_casper = ["casper-contract/wee_alloc", "casper-types"]
@@ -44,10 +44,10 @@ casper-types = { version = "^4.0.2", default-features = false, features = [], op
 radix-common = { version = "^1.3.0", default-features = false, features = [], optional = true }
 scrypto = { version = "^1.3.0", optional = true }
 sha3 = { version = "^0.10.8", default-features = false, features = ["asm"] }
-k256 = { version = "^0.13.4", default-features = false, features = [], optional = true }
+k256 = { version = "^0.13.4", default-features = false, features = ["ecdsa"], optional = true }
 secp256k1 = { version = "^0.29.1", default-features = false, features = [], optional = true }
 hex = { version = "^0.4.3", default-features = false, features = [], optional = true }
-primitive-types = { version = "^0.13.1", optional = true }
+primitive-types = { version = "^0.13.1", default-features = false }
 anchor-lang = { git = "https://github.com/coral-xyz/anchor", rev = "a7a23eea", optional = true, default-features = false }
