+++ src/core/mod.rs
@@ -5,6 +5,9 @@ pub mod processor_result;
 mod aggregator;
 mod validator;
 
+pub use processor::process_payload;
+pub use processor_result::ProcessorResult;
+
 #[cfg(feature = "helpers")]
 #[cfg(test)]
 mod test_helpers;
