+++ src/lib.rs
@@ -13,7 +13,44 @@ mod protocol;
 mod types;
 mod utils;
 
+use core::config::Config;
+use network::{Environment, StdEnv};
+
+pub use crypto::{CryptoError, DefaultCrypto, RecoverPublicKey};
 pub use types::{Bytes, FeedId, SignerAddress, TimestampMillis, Value};
+
+/// Configuration for the redstone protocol.
+/// Pluggable with custom environments and possible speciallized crypto operations.
+
+pub trait RedstoneConfig {
+    /// Crypto operations needed for address recovery.
+    type RecoverPublicKey: RecoverPublicKey;
+    /// Environment in which we execute. Provides logging etc
+    type Environment: Environment;
+
+    /// returns config for payload decoding and validation.
+    fn config(&self) -> &Config;
+}
+
+/// Standard nonspecialized implementation of the RedstoneConfig.
+/// See [crate::crypto::DefaultCrypto] for more information about crypto ops used.
+/// Constructuble from the `[crate::core::Config]`
+pub struct StdConfig(Config);
+
+impl From<Config> for StdConfig {
+    fn from(value: Config) -> Self {
+        Self(value)
+    }
+}
+
+impl RedstoneConfig for StdConfig {
+    type RecoverPublicKey = DefaultCrypto;
+    type Environment = StdEnv;
+
+    fn config(&self) -> &Config {
+        &self.0
+    }
+}
 
 #[cfg(feature = "helpers")]
 pub mod helpers;
