+++ src/protocol/data_point.rs
@@ -1,12 +1,6 @@
 use crate::{
-    network::{
-        as_str::{AsAsciiStr, AsHexStr},
-        assert::Assert,
-        error::Error,
-    },
-    protocol::constants::DATA_FEED_ID_BS,
+    network::as_str::{AsAsciiStr, AsHexStr},
     types::Value,
-    utils::trim::Trim,
     FeedId,
 };
 use std::fmt::{Debug, Formatter};
@@ -15,33 +9,6 @@ use std::fmt::{Debug, Formatter};
 pub(crate) struct DataPoint {
     pub(crate) feed_id: FeedId,
     pub(crate) value: Value,
-}
-
-pub(crate) fn trim_data_points(
-    payload: &mut Vec<u8>,
-    count: usize,
-    value_size: usize,
-) -> Vec<DataPoint> {
-    count.assert_or_revert(|&count| count == 1, |&count| Error::SizeNotSupported(count));
-
-    let mut data_points = Vec::new();
-
-    for _ in 0..count {
-        let data_point = trim_data_point(payload, value_size);
-        data_points.push(data_point);
-    }
-
-    data_points
-}
-
-fn trim_data_point(payload: &mut Vec<u8>, value_size: usize) -> DataPoint {
-    let value: Vec<_> = payload.trim_end(value_size);
-    let feed_id = payload.trim_end(DATA_FEED_ID_BS);
-
-    DataPoint {
-        value: value.into(),
-        feed_id,
-    }
 }
 
 impl Debug for DataPoint {
@@ -54,106 +21,5 @@ impl Debug for DataPoint {
             self.feed_id.as_hex_str(),
             self.value,
         )
-    }
-}
-
-#[cfg(feature = "helpers")]
-#[cfg(test)]
-mod tests {
-    use crate::{
-        helpers::hex::hex_to_bytes,
-        protocol::{
-            constants::DATA_FEED_ID_BS,
-            data_point::{trim_data_point, trim_data_points, DataPoint},
-        },
-        types::VALUE_SIZE,
-        Value,
-    };
-    use std::ops::Shr;
-
-    #[cfg(target_arch = "wasm32")]
-    use wasm_bindgen_test::wasm_bindgen_test as test;
-
-    const DATA_POINT_BYTES_TAIL: &str = "4554480000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000360cafc94e";
-    const VALUE: u128 = 232141080910;
-
-    #[test]
-    fn test_trim_data_points() {
-        let mut bytes = hex_to_bytes(DATA_POINT_BYTES_TAIL.into());
-        let result = trim_data_points(&mut bytes, 1, 32);
-
-        assert_eq!(result.len(), 1);
-
-        verify_rest_and_result(
-            DATA_POINT_BYTES_TAIL,
-            32,
-            VALUE.into(),
-            bytes,
-            result[0].clone(),
-        )
-    }
-
-    #[should_panic(expected = "Size not supported: 0")]
-    #[test]
-    fn test_trim_zero_data_points() {
-        trim_data_points(&mut hex_to_bytes(DATA_POINT_BYTES_TAIL.into()), 0, 32);
-    }
-
-    #[should_panic(expected = "Size not supported: 2")]
-    #[test]
-    fn test_trim_two_data_points() {
-        trim_data_points(&mut hex_to_bytes(DATA_POINT_BYTES_TAIL.into()), 2, 32);
-    }
-
-    #[test]
-    fn test_trim_data_point() {
-        test_trim_data_point_of(DATA_POINT_BYTES_TAIL, 32, VALUE.into());
-    }
-
-    #[test]
-    fn test_trim_data_point_with_prefix() {
-        test_trim_data_point_of(
-            &("a2a812f3ee1b".to_owned() + DATA_POINT_BYTES_TAIL),
-            32,
-            VALUE.into(),
-        );
-    }
-
-    #[test]
-    fn test_trim_data_point_other_lengths() {
-        for i in 1..VALUE_SIZE {
-            test_trim_data_point_of(
-                &DATA_POINT_BYTES_TAIL[..DATA_POINT_BYTES_TAIL.len() - 2 * i],
-                32 - i,
-                Value::from_u256(primitive_types::U256::from(VALUE).shr(8 * i as u32)),
-            );
-        }
-    }
-
-    fn test_trim_data_point_of(value: &str, size: usize, expected_value: Value) {
-        let mut bytes = hex_to_bytes(value.into());
-        let result = trim_data_point(&mut bytes, size);
-
-        verify_rest_and_result(value, size, expected_value, bytes, result);
-    }
-
-    fn verify_rest_and_result(
-        value: &str,
-        size: usize,
-        expected_value: Value,
-        rest: Vec<u8>,
-        result: DataPoint,
-    ) {
-        assert_eq!(
-            rest,
-            hex_to_bytes(value[..value.len() - 2 * (size + DATA_FEED_ID_BS)].into())
-        );
-
-        let data_point = DataPoint {
-            value: expected_value,
-            feed_id: hex_to_bytes(DATA_POINT_BYTES_TAIL[..6].to_string()).into(),
-        };
-
-        assert_eq!(result, data_point);
     }
 }
