+++ src/protocol/mod.rs
@@ -1,5 +1,9 @@
 pub(crate) mod constants;
 pub(crate) mod data_package;
 pub(crate) mod data_point;
-pub(crate) mod marker;
+mod marker;
 pub(crate) mod payload;
+
+mod payload_decoder;
+
+pub use payload_decoder::PayloadDecoder;
